+++ ErsatzTV.Core/FFmpeg/FFmpegLibraryProcessService.cs
@@ -150,7 +150,12 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
             playbackSettings.AudioBufferSize,
             playbackSettings.AudioSampleRate,
             videoPath == audioPath ? playbackSettings.AudioDuration : Option<TimeSpan>.None,
-            playbackSettings.NormalizeLoudness);
+            playbackSettings.NormalizeLoudnessMode switch
+            {
+                NormalizeLoudnessMode.LoudNorm => AudioFilter.LoudNorm,
+                NormalizeLoudnessMode.DynAudNorm => AudioFilter.DynAudNorm,
+                _ => AudioFilter.None
+            });
 
         // don't log generated images, or hls direct, which are expected to have unknown format
         bool isUnknownPixelFormatExpected =
@@ -301,15 +306,26 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
             ? Path.Combine(FileSystemLayout.TranscodeFolder, channel.Number, "live%06d.ts")
             : Option<string>.None;
 
+        FrameSize scaledSize = ffmpegVideoStream.SquarePixelFrameSize(
+            new FrameSize(channel.FFmpegProfile.Resolution.Width, channel.FFmpegProfile.Resolution.Height));
+
+        var paddedSize = new FrameSize(
+            channel.FFmpegProfile.Resolution.Width,
+            channel.FFmpegProfile.Resolution.Height);
+
+        if (channel.FFmpegProfile.ScalingBehavior is ScalingBehavior.Stretch)
+        {
+            scaledSize = paddedSize;
+        }
+        
         var desiredState = new FrameState(
             playbackSettings.RealtimeOutput,
             fillerKind == FillerKind.Fallback,
             videoFormat,
             Optional(videoStream.Profile),
             Optional(playbackSettings.PixelFormat),
-            ffmpegVideoStream.SquarePixelFrameSize(
-                new FrameSize(channel.FFmpegProfile.Resolution.Width, channel.FFmpegProfile.Resolution.Height)),
-            new FrameSize(channel.FFmpegProfile.Resolution.Width, channel.FFmpegProfile.Resolution.Height),
+            scaledSize,
+            paddedSize,
             false,
             playbackSettings.FrameRate,
             playbackSettings.VideoBitrate,
@@ -403,7 +419,7 @@ public class FFmpegLibraryProcessService : IFFmpegProcessService
             playbackSettings.AudioBufferSize,
             playbackSettings.AudioSampleRate,
             Option<TimeSpan>.None,
-            false);
+            AudioFilter.None);
 
         var desiredState = new FrameState(
             playbackSettings.RealtimeOutput,