+++ ErsatzTV.FFmpeg/AudioFilter.cs
@@ -1,0 +1,8 @@
+namespace ErsatzTV.FFmpeg;
+
+public enum AudioFilter
+{
+    None = 0,
+    LoudNorm = 1,
+    DynAudNorm = 2
+}