+++ ErsatzTV.Infrastructure.Sqlite/Migrations/20230921071647_Add_FFmpegProfileNormalizeLoudnessMode.cs
@@ -1,0 +1,31 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class Add_FFmpegProfileNormalizeLoudnessMode : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "NormalizeLoudnessMode",
+                table: "FFmpegProfile",
+                type: "INTEGER",
+                nullable: false,
+                defaultValue: 0);
+            
+            migrationBuilder.Sql("update FFmpegProfile set NormalizeLoudnessMode = 1 where NormalizeLoudness = 1");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "NormalizeLoudnessMode",
+                table: "FFmpegProfile");
+        }
+    }
+}